requirements.txt
fastapi==0.110.2
httpx==0.23.3
firebase==4.0.1
loguru==0.6.0
python-decouple==3.8
uvicorn==0.20.0
Inject==4.3.1
google-cloud-secret-manager==2.7.0
firebase-admin==6.5.0
pytest-asyncio==0.21.0
pytest==7.2.2
mysql-connector-python==8.3.0
pymysql==1.1.0
pytz==2024.1